// src/blocks/CityDetails/types.ts
export interface CityDetailsProps {
	cityName: string;
	country: string;
	icon: string;
	description: string;
	iconType: string;
	temperature: number;
	minTemp: number;
	maxTemp: number;
	humidity: number;
	pressure: number;
	feelsLike: number;
	visibility: number;
	windSpeed: number;
	windDegree: number;
	clouds: number;
	sunrise: string;
	sunset: string;
}